Fix failing build on aarch64-linux-android
The types of arguments `hostlen` and `sevlen` on the libc function getnameinfo
differs between android anlinux. On linux they are `::socklen_t` which is
defined as a u32. Onandroid they are `::size_t` which is a usize. The
explicit cast to an u32 here make the build fail on android 64 bits.
Using `as _` instead of `as u32` makes the compiler choose the right
type for us at compile time