pass NI_NUMERICSERV to getnameinfo in lookup_addr
`lookup_addr` only returns the hostname, so there is no reason for us to
have `libnss` and friends look up the service name.

Currently, tight loops of `lookup_addr` result in a significant amount
of CPU time being spent in `libnss_files.so` to parse service names, as
`/etc/services` is generally tens of thousands of lines unlike `/etc/hosts`.

As an example, here is a redacted perf report before and after this commit:
https://gist.github.com/rahulg/e9440f65308dd2a9bdb8cfd815b95d0f